requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
torch>=2.1.0
torchvision>=0.16.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
redis>=5.0.1
rq>=1.16.0
prometheus-client>=0.20.0
httpx>=0.26.0
Pillow>=10.2.0
click>=8.1.7
structlog>=24.1.0
numpy>=1.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
